requirements.txt
# 基础科学计算库
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 神经影像处理
nibabel>=3.2.0
nilearn>=0.9.0

# 大脑可视化
brainspace>=0.1.2
matplotlib>=3.5.0
seaborn>=0.11.0

# 图处理
networkx>=2.6.0
python-louvain>=0.15

# 信号处理
scikit-learn>=1.0.0
statsmodels>=0.13.0

# 配置文件处理
pyyaml>=5.4.0

# 数据可视化
plotly>=5.3.0

# 工具库
tqdm>=4.62.0
h5py>=3.6.0

# 开发工具
pytest>=6.2.0
jupyter>=1.0.0
